+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- feat: add `code` and `cwd` parameters to `PythonInterpreterTool` (#490)
+- feat: add `stdin` parameter to `PythonInterpreterTool` (#487)
+
 ### Changed
 
 - refactor: rip out disable_model_invocation from LLMAgent (#456)